Switching from interface to generic api class
