+++ config/params.ini
@@ -2,7 +2,7 @@
 [general]
 name=Botty
 monitor=0
-message_api_type=discord_basic
+message_api_type=generic_api
 custom_message_hook=
 logg_lvl=debug
 max_game_length_s=380
