+++ src/api/__init__.py
@@ -1,1 +1,2 @@
-from .i_api import IApi
+from .generic_api import GenericApi
+from .discord_embeds import DiscordEmbeds
